Fix not retrying tile downloads.
Move all logic deciding whether tile has been loaded to
TMSCachedTileLoader. If loaded is set to false, then it means that even
if error is set, the error is transient and we should try to download
tile again.

Create logic in TMSCachedTileLoader deciding when we are facing
transient errors:
* when HTTP status code is greater than 500
* when no cache attributes are set (assume lost job or exception before
getting http response - i.e. no Internet connection, but also - typo in
hostname)